Setup effects structure.

+++ app/src/main/java/com/github/kutyrev/poemmaster/model/PoemHeader.kt
@@ -1,3 +1,3 @@
 package com.github.kutyrev.poemmaster.model
 
-data class PoemHeader(val id: Int, val name: String)
+data class PoemHeader(val id: Long, val name: String)
